+++ jslint_ci.sh
@@ -1944,7 +1944,6 @@ function globExclude({
       /[\[\]]/g
     ), "\\$&");
     strRegex = strRegex.replace((
-      // ignore [-/]
       /[$()*+.?\[\\\]\^{|}]/g
     ), "\\$&");
     strRegex = strRegex.replace((