+++ test.mjs
@@ -861,8 +861,9 @@ jstestDescribe((
             ],
             var: [
 
-// PR-363 - Bugfix - add test against false-warning
-// <uninitialized 'bb'> in code '/*jslint node*/\nlet {aa:bb} = {}; bb();'
+// PR-363 - Bugfix
+// Add test against false-warning <uninitialized 'bb'> in code
+// '/*jslint node*/\nlet {aa:bb} = {}; bb();'.
 
                 "/*jslint node*/\n",
                 ""
@@ -921,6 +922,9 @@ jstestDescribe((
         ], [
             "debugger;", {devel: true}, []
         ], [
+
+// PR-404 - Alias "evil" to jslint-directive "eval" for backwards-compat.
+
             "new Function();\neval();", {eval: true, evil: true}, []
         ], [
             (
@@ -943,12 +947,18 @@ jstestDescribe((
         ], [
             "/".repeat(100), {long: true}, []
         ], [
+
+// PR-404 - Alias "nomen" to jslint-directive "name" for backwards-compat.
+
             "let aa = aa._;", {name: true, nomen: true}, []
         ], [
             "require();", {node: true}, []
         ], [
             "let aa = 'aa';", {single: true}, []
         ], [
+
+// PR-404 - Add new directive "subscript" to play nice with Google Closure.
+
             "aa[\"aa\"] = 1;", {subscript: true}, ["aa"]
         ], [
             "", {test_internal_error: true}, []